# swagger.yml
openapi: "3.0.0"
info:
  description: "This is a sample server for Screenz test project. For this sample, you can use the user/password pairs `admin/admin` and `client/client` to test the authorization."
  version: "1.0.0"
  title: "Screenz Test Project"
  contact:
    email: "shprota@gmail.com"
servers:
  - url: http://screenz.shprota.com/api
    description: Example cloud server
  - url: http://localhost:3000/api
    description: Locally running server
...
paths:
...
  /books:
    get:
      tags:
        - books
      summary: "List books"
      description: "Returns the list of books"
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArrayOfBooks'
        401:
          description: "Not logged in"
        403:
          description: "Admin permission required"
        500:
          description: "Any other error"

    post:
      tags:
        - books
      summary: Create new book
      description: Only admin is allowed
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CreateBook'
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBook'
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        401:
          description: "Not logged in"
        403:
          description: "Admin permission required"
        500:
          description: "Any other error"
  /books/{id}:
    put:
      tags:
        - books
      summary: "Update book"
      description: "Updates the book with specified ID"
      parameters:
        - name: id
          in: path
          description: Book ID
          required: true
          schema:
            type: string
            format: int64
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CreateBook'
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBook'
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        401:
          description: "Not logged in"
        403:
          description: "Admin permission required"
        400:
          description: "Book not found"
        500:
          description: "Any other error"
    delete:
      tags:
        - books
      summary: "Delete book"
      description: "Deletes the book with specified ID"
      parameters:
        - name: id
          in: path
          description: Book ID
          required: true
          schema:
            type: string
            format: int64
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        401:
          description: "Not logged in"
        403:
          description: "Admin permission required"
        400:
          description: "Book not found"
        500:
          description: "Any other error"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: connect.sid
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Success:
      type: object
      properties:
        success:
          type: boolean
          example: true
...
    Review:
      type: object
      properties:
        _id:
          type: string
          format: int64
          example: 5c596cf84c406319778f6bcc
        user:
          type: string
          format: int64
          example: 5c596cf84c406319778f6bcc
        rating:
          type: integer
          minimum: 1
          maximum: 10
          example: 7
        status:
          type: string
          enum: ['Submitted', 'Rejected', 'Accepted']
        description:
          type: string
          example: I loved reading this book to my grandchildren
...
    Book:
      type: object
      properties:
        _id:
          type: string
          format: int64
          example: 5c596cf84c406319778f6bcc
        name:
          type: string
          example: "The Catcher in the Rye"
        genre:
          type: string
          format: int64
          example: 5c596cf84c406319778f6bcc
        description:
          type: string
          example: The Catcher in the Rye is a story by J. D. Salinger, partially published in serial form in 1945–1946 and as a novel in 1951. A classic novel originally published for adults, it has since become popular among adolescent readers for its themes of angst and alienation.
        reviews:
          type: array
          items:
            $ref: '#/components/schemas/Review'
    CreateBook:
      type: object
      properties:
        name:
          type: string
          example: "The Catcher in the Rye"
        genre:
          type: string
          format: int64
          example: 5c596cf84c406319778f6bcc
        description:
          type: string
          example: The Catcher in the Rye is a story by J. D. Salinger, partially published in serial form in 1945–1946 and as a novel in 1951. A classic novel originally published for adults, it has since become popular among adolescent readers for its themes of angst and alienation.
    ArrayOfBooks:
      type: array
      items:
        $ref: '#/components/schemas/Book'
security:
  - cookieAuth: []
  - bearerAuth: []